+++ src/language_features/inlay_hints.rs
@@ -1,7 +1,7 @@
 use itertools::Itertools;
 use lsp_types::{
-    request::{InlayHintRequest, Request},
-    InlayHint, InlayHintLabel, InlayHintParams, Position, Range, TextDocumentIdentifier, Url,
+    request::InlayHintRequest, InlayHint, InlayHintLabel, InlayHintParams, Position, Range,
+    TextDocumentIdentifier, Url,
 };
 use serde::Deserialize;
 
@@ -15,29 +15,6 @@ use crate::{
 #[derive(Debug, PartialEq, Clone, Default, Serialize, Deserialize)]
 struct InlayHintsOptions {
     buf_line_count: u32,
-}
-
-pub enum ExperimentalInlayHintRequest {}
-
-// For now, rust analyzer implements this as experimental/inlayHint
-impl Request for ExperimentalInlayHintRequest {
-    type Params = InlayHintParams;
-    type Result = Option<Vec<InlayHint>>;
-    const METHOD: &'static str = "experimental/inlayHints";
-}
-
-pub fn experimental_inlay_hints(meta: EditorMeta, params: EditorParams, ctx: &mut Context) {
-    let params = InlayHintsOptions::deserialize(params).unwrap();
-    let req_params = InlayHintParams {
-        work_done_progress_params: Default::default(),
-        text_document: TextDocumentIdentifier {
-            uri: Url::from_file_path(&meta.buffile).unwrap(),
-        },
-        range: Range::new(Position::new(0, 0), Position::new(params.buf_line_count, 0)),
-    };
-    ctx.call::<ExperimentalInlayHintRequest, _>(meta, req_params, move |ctx, meta, response| {
-        inlay_hints_response(meta, response.unwrap_or_default(), ctx)
-    });
 }
 
 pub fn inlay_hints(meta: EditorMeta, params: EditorParams, ctx: &mut Context) {